Add release tooling, debian packaging, and README with usage preview
- Add release.sh for automated multi-arch builds and packaging (deb, rpm, apk, sh)
- Add debian/ source packaging for Launchpad PPA uploads
- Fix 32-bit build (int32→int64 cast in platform_linux.go)
- Update README with sample output, PPA install instructions, and docs
- Update .gitignore to exclude vendor/ and release/

+++ platform_linux.go
@@ -152,8 +152,8 @@ func restoreMetadata(path string, orig os.FileInfo) error {
 	}
 
 	// Timestamps.
-	atime := time.Unix(stat.Atim.Sec, stat.Atim.Nsec)
-	mtime := time.Unix(stat.Mtim.Sec, stat.Mtim.Nsec)
+	atime := time.Unix(int64(stat.Atim.Sec), int64(stat.Atim.Nsec))
+	mtime := time.Unix(int64(stat.Mtim.Sec), int64(stat.Mtim.Nsec))
 	if err := os.Chtimes(path, atime, mtime); err != nil {
 		return fmt.Errorf("chtimes: %w", err)
 	}
